+++ .travis.yml
@@ -1,14 +1,17 @@
 language: python
 python:
-- "3.6"
+  - "3.6"
 
 install:
-- pip install -r requirements.txt
+  - pip install -r requirements.txt
 before_script:
-- psql -c 'create database STORE_DB;' -U postgres
+  - curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
+  - chmod +x ./cc-test-reporter
+  - ./cc-test-reporter before-build
+  - psql -c 'create database STORE_DB;' -U postgres
 script:
-- pytest --cov=storemanager/api/v2
+  - pytest --cov=storemanager/api/v2
 
 after_success:
-- coveralls
-- codeclimate-test-reporter
+  - coveralls
+  - ./cc-test-reporter after-build --exit-code $TRAVIS_TEST_RESULT
