+++ storemanager/api/v2/database/queries.py
@@ -1,4 +1,7 @@
-"""The Query module contains all the necessary queries that are used in the application"""
+"""
+The Query module contains all the necessary queries
+that are used in the application
+"""
 
 CREATE_TABLE_USERS = """
     CREATE TABLE USERS (
@@ -44,62 +47,64 @@ CREATE_TABLE_SALE_ITEMS = """
     );"""
 
 DROP_ALL_TABLES = """
-    DROP TABLE IF EXISTS USERS, PRODUCTS, CATEGORIES, SALE_RECORDS, SALE_RECORD_ITEMS;"""
+DROP TABLE IF EXISTS USERS, PRODUCTS, CATEGORIES,
+SALE_RECORDS, SALE_RECORD_ITEMS;"""
 
 CREATE_CATEGORY = """
-    INSERT INTO categories(name, description) 
-    VALUES(%s, %s) 
+    INSERT INTO categories(name, description)
+    VALUES(%s, %s)
     RETURNING id, name, description;"""
 
 GET_CATEGORY = """
-    SELECT id, name, description 
-    FROM categories 
+    SELECT id, name, description
+    FROM categories
     WHERE id = %s"""
 
 GET_CATEGORY_BY_NAME = """
-    SELECT id, name, description 
-    FROM categories 
+    SELECT id, name, description
+    FROM categories
     WHERE name = %s"""
 
 DELETE_CATEGORY = """
-    DELETE FROM categories 
+    DELETE FROM categories
     WHERE id = %s"""
 
 UPDATE_CATEGORY = """
-    UPDATE categories 
-    SET name = %s, description = %s 
+    UPDATE categories
+    SET name = %s, description = %s
     WHERE id = %s"""
 
 GET_ALL_CATEGORIES = """
-    SELECT id, name, description 
-    FROM categories 
+    SELECT id, name, description
+    FROM categories
     ORDER BY id"""
 
 CREATE_PRODUCT = """
-    INSERT INTO products(name, price, stock, stockmin, description, category) 
-    VALUES(%s, %s, %s,%s, %s, %s) 
+    INSERT INTO products(name, price, stock, stockmin, description, category)
+    VALUES(%s, %s, %s,%s, %s, %s)
     RETURNING id, name, description, price, stock, stockmin, category;"""
 
 GET_PRODUCT = """
-    SELECT * FROM products 
+    SELECT * FROM products
     WHERE id = %s"""
 
 GET_PRODUCT_BY_NAME = """
-    SELECT id, name, price, stock, stockmin 
-    FROM products 
+    SELECT id, name, price, stock, stockmin
+    FROM products
     WHERE name = %s"""
 
 DELETE_PRODUCT = """
-    DELETE FROM products 
+    DELETE FROM products
     WHERE id = %s"""
 
 UPDATE_PRODUCT = """
-    UPDATE products 
-    SET name = %s, description = %s, price = %s, stock = %s, stockmin = %s, category = %s 
+    UPDATE products
+    SET name = %s, description = %s, price = %s,
+    stock = %s, stockmin = %s, category = %s
     WHERE id = %s;"""
 
 UPDATE_PRODUCT_ON_SALE = """
-    UPDATE products 
+    UPDATE products
     SET stock = %s
     WHERE id = %s;"""
 
@@ -107,56 +112,56 @@ GET_ALL_PRODUCTS = """
     SELECT * FROM products ORDER BY id"""
 
 CREATE_SALE = """
-    INSERT INTO sale_records(items, total) 
-    VALUES(%s, %s) 
+    INSERT INTO sale_records(items, total)
+    VALUES(%s, %s)
     RETURNING id, items, total;"""
 
 GET_SALE = """
-    SELECT id, items, total 
-    FROM sale_records 
+    SELECT id, items, total
+    FROM sale_records
     WHERE id = %s"""
 
 GET_ALL_SALES = """
-    SELECT id, items, total 
-    FROM sale_records 
+    SELECT id, items, total
+    FROM sale_records
     ORDER BY id"""
 
 CREATE_SALE_ITEM = """
-    INSERT INTO sale_record_items(product_name,price,quantity,total,sale_id) 
+    INSERT INTO sale_record_items(product_name,price,quantity,total,sale_id)
     VALUES(%s, %s, %s, %s, %s)"""
 # RETURNING product_id, price, quantity, total;"""
 
 GET_SALE_ITEMS = """
-    SELECT product_name, price, quantity, total 
-    FROM sale_record_items 
-    WHERE sale_id = %s 
+    SELECT product_name, price, quantity, total
+    FROM sale_record_items
+    WHERE sale_id = %s
     ORDER BY id"""
 
 CREATE_USER = """
-    INSERT INTO users(name, password,role) 
-    VALUES(%s, %s, %s) 
+    INSERT INTO users(name, password,role)
+    VALUES(%s, %s, %s)
     RETURNING id, name, role;"""
 
 GET_USER = """
-    SELECT id, name, role 
-    FROM users 
+    SELECT id, name, role
+    FROM users
     WHERE id = %s"""
 
 GET_USER_BY_NAME = """
-    SELECT id, name, password, role 
-    FROM users 
+    SELECT id, name, password, role
+    FROM users
     WHERE name = %s"""
 
 DELETE_USER = """
-    DELETE FROM users 
+    DELETE FROM users
     WHERE id = %s"""
 
 UPDATE_USER = """
-    UPDATE users 
-    SET name = %s, password = %s 
+    UPDATE users
+    SET name = %s, password = %s
     WHERE id = %s"""
 
 GET_ALL_USERS = """
-    SELECT id, name, role 
-    FROM users 
+    SELECT id, name, role
+    FROM users
     ORDER BY id"""
